backend/selection-service/pyproject.toml
[project]
name = "selection-service"
version = "0.1.0"
description = "Image area selection service"
requires-python = ">=3.9,<3.12"
dependencies = [
    "aibox-image-lib",
    "numpy>=2.0.2",
    "zmq>=0.0.0",
    "torch>=2.8.0",
    "transformers>=4.56.1",
    "pillow>=11.3.0",
    "opencv-python>=4.11.0.86",
    "torchvision>=0.23.0",
]

[tool.uv.sources]
aibox-image-lib = { path = "../aibox-image-lib", editable = true }

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=23.0.0"]

[project.scripts]
service = "selection_service.main:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
